# project_management/market_data_api_openapi.yaml
openapi: 3.0.3
info:
  title: Omnitrade Market Data REST API
  version: '1.0'
  description: API for accessing market data including symbols, ticker, orderbook, trades, and klines.
servers:
  - url: https://api.omnitrade.example.com/api/v1/market-data
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  parameters:
    SymbolParam:
      name: symbol
      in: query
      required: true
      schema:
        type: string
      description: Trading pair symbol, e.g., BTCUSDT
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
      description: Number of results to return
    IntervalParam:
      name: interval
      in: query
      required: true
      schema:
        type: string
      description: Candlestick interval (e.g., 1m, 5m, 1h)
    StartTimeParam:
      name: startTime
      in: query
      required: false
      schema:
        type: integer
      description: Start timestamp in milliseconds
    EndTimeParam:
      name: endTime
      in: query
      required: false
      schema:
        type: integer
      description: End timestamp in milliseconds
security:
  - ApiKeyAuth: []
paths:
  /symbols:
    get:
      summary: List all supported trading symbols
      responses:
        '200':
          description: List of symbols
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
              example: ["BTCUSDT", "ETHUSDT"]
        '500':
          description: Server error
      security: []
  /ticker:
    get:
      summary: Get latest ticker price for a symbol
      parameters:
        - $ref: '#/components/parameters/SymbolParam'
      responses:
        '200':
          description: Ticker data
          content:
            application/json:
              schema:
                type: object
                properties:
                  symbol:
                    type: string
                  price:
                    type: string
              example:
                symbol: BTCUSDT
                price: "50000.00"
        '400':
          description: Invalid request parameters
        '500':
          description: Server error
  /orderbook:
    get:
      summary: Get current order book snapshot
      parameters:
        - $ref: '#/components/parameters/SymbolParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Order book data
          content:
            application/json:
              schema:
                type: object
                properties:
                  bids:
                    type: array
                    items:
                      type: array
                      items:
                        type: string
                  asks:
                    type: array
                    items:
                      type: array
                      items:
                        type: string
              example:
                bids: [["50000.00", "1.2"], ["49950.00", "0.5"]]
                asks: [["50010.00", "0.8"], ["50020.00", "1.0"]]
        '400':
          description: Invalid request parameters
        '500':
          description: Server error
  /trades:
    get:
      summary: Get recent trades
      parameters:
        - $ref: '#/components/parameters/SymbolParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: List of recent trades
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    price:
                      type: string
                    quantity:
                      type: string
                    timestamp:
                      type: integer
              example:
                - price: "50000.00"
                  quantity: "0.1"
                  timestamp: 1617181723000
                - price: "50010.00"
                  quantity: "0.2"
                  timestamp: 1617181724000
        '400':
          description: Invalid request parameters
        '500':
          description: Server error
  /klines:
    get:
      summary: Get historical candlestick data
      parameters:
        - $ref: '#/components/parameters/SymbolParam'
        - $ref: '#/components/parameters/IntervalParam'
        - $ref: '#/components/parameters/StartTimeParam'
        - $ref: '#/components/parameters/EndTimeParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: List of candlesticks
          content:
            application/json:
              schema:
                type: array
                items:
                  type: array
                  items:
                    oneOf:
                      - type: integer
                      - type: string
              example:
                - [1617181723000, "50000.00", "50100.00", "49900.00", "50050.00", "100.0"]
                - [1617181783000, "50050.00", "50200.00", "50000.00", "50150.00", "150.0"]
        '400':
          description: Invalid request parameters
        '500':
          description: Server error